Added a silly MyInputSystem + Tests!
The problem I have struggled with forever is that polling for input yields a barrage of
	UP
	UP
	UP
	UP
	UP
	UP
	DOWN
	UP
	DOWN
	UP
	LEFT
	DOWN
	...

but what I _want_ is:
	"Up was pressed"
	"Down was pressed"
	"Left was pressed"
	"Up was released"
	...

I learned about Amethyst's Event Channels and thought "I can use that".
So now we have a sytem that reads Input events and spawns custom MyInput events which other systems can listen to.
It doesn't do that "Up was released" but it can pretty easily.

This also includes a boilerplate system to listen to said custom events, which will become the "movement" system.

Finally it incudes _tests_ for the MyInputSystem which is sick!
I'm _so_ lazy! I'm so glad I don't need to manually test stuff.