Wow well do I have a story for you.
So I was encountering a bug where every time I reloaded a level the "door" would disapear and then rea-appear when I reloaded again.

This is weird behavior, and based on what I was researching about game programming, of course I jumped to assuming that this was an asset loading timing problem.
After all, I was re-loading the spritesheet and texture map every time I loaded the level.
Clearly that was _so_ inefficient that it caused this sprite not to load.

(I don't know if you can detect the sarcasm in that last sentence, but this was a very silly leap to make)

So I spent all night, and a morning or two, trying to streamline my asset loading.
	To be fair, they're all necessary changes.
	I should be doing much more _waiting_ in the `Loading` state than I am -- or at least _checking_ that what we'll need in `Menu` and `Level` are ready to be used.
I got one attempt down, but it got too hairy for me to keep working on the re-write.
Then I got a second re-write where I was _just_ loading the spritesheet in the `Loading` state.

I finally got the new re-write to work and got the game loading and _womp womp_ the exact same behavior was still there.

Then it dawned on me that I have seen this problem before.
The floor tiles sometimes cover other sprites because all sprites have default z translation.

In two minutes I set the z translation for tiles to `-10.0`, and the *poof* the problem was gone.

So no, loading a 256x256 sprite sheet, in a hilariously tiny game, every few seconds when a level starts, was not causing a heisenbug.

So what did I learn?
Don't forget your z-translations.
And maybe make your code more refactorable.